rojcad::types

Enum ReplToViewer

pub enum ReplToViewer {
    FitToBounds {
        center: DVec3,
        radius: f64,
        keep_angle: bool,
    },
    SetViewAngles {
        yaw: f64,
        pitch: f64,
        distance: Option<f64>,
    },
    SetWindowSize {
        width: u32,
        height: u32,
    },
    SetFullscreen(bool),
    SetMaximized(bool),
}

Commands sent from the REPL thread to the viewer thread. The viewer polls these each frame via an mpsc receiver.

Variants

FitToBounds

Fit the camera to frame a bounding sphere.

Fields

center: DVec3

radius: f64

keep_angle: bool

SetViewAngles

Set camera to specific yaw/pitch angles, optionally with a distance.

Fields

yaw: f64

pitch: f64

distance: Option<f64>

SetWindowSize

Resize the viewer window to the given logical dimensions.

Fields

width: u32

height: u32

SetFullscreen(bool)

Enter or exit fullscreen mode.

SetMaximized(bool)

Enter or exit maximized state.